+++ horse64/valuecontentstruct.c
@@ -107,7 +107,7 @@ int valuecontent_SetStringU32(
         gcstr->str_val.s, s,
         sizeof(*s) * slen
     );
-    assert(gcstr->str_val.len == (unsigned)slen);
+    assert(gcstr->str_val.len == (uint64_t)slen);
     assert(gcstr->str_val.letterlen == 0);
     gcstr->type = H64GCVALUETYPE_STRING;
     return 1;
